Enable bitmap frame allocator on i686 (#919)
* Add x86::addr::PhysFrame

* Remove useless conversions

* Use bitfield for addr

* Use PhysFrame in Bitmap

* Remove unused code

* Enable bitmap on i686

* Simplify code with frame number

* Add tests

* Skip unaddressable regions

* Increase MemoryRegion capacity

* Refactor mapping

* Clip regions below 4 GB

* Reserve the first page

* Use bytes for the bitmap

* Stop printing memory holes

* Use aligned regions

* Remove TODO comment about alignment

* Enable bitmap

* Rename InterruptFrame to InterruptRegisters

* Rename Registers to SyscallRegisters

* Rename PhysFrame to Frame

* Add comments

* Rename tests

* Update changelog

+++ src/sys/mem/phys.rs
@@ -21,7 +21,7 @@ impl PhysBuf {
     fn from(vec: Vec<u8>) -> Self {
         let buffer_end = vec.len() - 1;
         let memory_end = phys_addr(&vec[buffer_end]) - phys_addr(&vec[0]);
-        if buffer_end == memory_end as usize {
+        if buffer_end == memory_end {
             Self {
                 buf: Arc::new(Mutex::new(vec)),
             }
